requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
requests
qdrant-client
fastembed
google-genai
